requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
aiohttp==3.7.3
    # via discord.py
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
cffi==1.14.4
    # via pynacl
chardet==3.0.4
    # via aiohttp
discord.py[voice]==1.6.0
    # via -r requirements/requirements.in
idna==3.1
    # via yarl
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
pycparser==2.20
    # via cffi
pynacl==1.4.0
    # via discord.py
python-dotenv==0.15.0
    # via -r requirements/requirements.in
six==1.15.0
    # via pynacl
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   yarl
yarl==1.6.3
    # via aiohttp
